# Conflicts:
#	prj/Ren_U2A16_373_CDC_1MJP/PE3VM3/App.gpj
#	src/PE3VM3/App/Central/esw/EthSW_Task/EthSW_Task.c
#	src/PE3VM3/App/Central/esw/EthSwt/cfg/EthSwt_SWIC_initRegCommon.h
#	src/PE3VM3/App/Central/esw/EthSwt/inc/EthSwt_SWIC_Def.h
#	src/PE3VM3/App/Central/esw/EthSwt/inc/EthSwt_SWIC_Org.h
#	src/PE3VM3/App/Central/esw/EthSwt/inc/EthSwt_SWIC_Reg.h
#	src/PE3VM3/App/Central/esw/EthSwt/src/EthSwt_SWIC.c
#	src/PE3VM3/App/Central/esw/EthSwt/src/EthSwt_SWIC_Org.c
#	src/PE3VM3/App/Central/esw/EthSwt/src/EthSwt_SWIC_Reg.c
#	src/PE3VM3/App/Central/esw/EthSwt/src/EthSwt_SWIC_SpiS1.c
#	src/PE3VM3/App/Central/esw/inc/EthSwt_SWIC.h

+++ src/PE3VM3/App/Central/esw/EthSwt/inc/EthSwt_SWIC_Pwr.h
@@ -10,7 +10,7 @@
 #include <Port_Cfg.h>
 #include <Port.h>
 /* -------------------------------------------------------------------------- */
-#define ETHSWT_SWIC_PWR_GET_SAIL_RESOUT_N       (uint8)Dio_ReadChannel(DIO_ID_PORT8_CH8)
+#define ETHSWT_SWIC_PWR_GET_SAIL_RESOUT_N       Dio_ReadChannel(DIO_ID_PORT8_CH8)
 
 #define ETHSWT_SWIC_PWR_ETHER_PWR_EN_HIGH       Dio_WriteChannel(DIO_ID_APORT4_CH11, STD_HIGH)
 #define ETHSWT_SWIC_PWR_ETHER_PWR_EN_LOW        Dio_WriteChannel(DIO_ID_APORT4_CH11, STD_LOW)
